net/cnxk: fix build on Ubuntu 24.04
[ upstream commit b9799fb5e7a38c824c91b88d3c89250d23c783e6 ]

Due to implicit unsigned to signed integer conversion, actual value gets
wrapped and becomes higher than its size.

Bugzilla ID: 1513
Fixes: 03b152389fb1 ("net/cnxk: add option to enable custom inbound SA")
Fixes: 7df4ead35436 ("net/cnxk: support parsing custom SA action")
Fixes: 47cca253d605 ("net/cnxk: support Rx inject")

+++ drivers/net/cnxk/cnxk_ethdev_devargs.c
@@ -269,7 +269,7 @@ cnxk_ethdev_parse_devargs(struct rte_devargs *devargs, struct cnxk_eth_dev *dev)
 	uint16_t custom_sa_act = 0;
 	struct rte_kvargs *kvlist;
 	uint16_t no_inl_dev = 0;
-	uint8_t lock_rx_ctx = 0;
+	uint16_t lock_rx_ctx = 0;
 
 	memset(&sdp_chan, 0, sizeof(sdp_chan));
 	memset(&pre_l2_info, 0, sizeof(struct flow_pre_l2_size_info));
